# backend/openapi.yaml
openapi: 3.0.3
info:
  title: Beat Portal API
  description: |
    FastAPI backend for Beat Portal - A DJ music library organizer with AI-powered metadata enhancement.
    
    ## Features
    - Offline-first music library management
    - Advanced metadata analysis (BPM, Key, Genre, Mood)
    - AI-powered metadata enhancement
    - Track organization and playlist management
  version: 1.0.0
  contact:
    name: Beat Portal Support
    email: support@beatportal.dev
servers:
  - url: http://localhost:8000/api/v1
    description: Local development server
  - url: http://localhost:8000
    description: Local base server
...
paths:
...
  /tracks:
    get:
      tags:
        - Tracks
      summary: List Tracks
      description: Get a paginated list of tracks with optional filtering and sorting
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          schema:
            type: integer
            default: 50
            maximum: 100
        - name: search
          in: query
          description: Search query for title, artist, or genre
          schema:
            type: string
        - name: genre
          in: query
          schema:
            type: string
        - name: mood
          in: query
          schema:
            type: string
        - name: bpm_min
          in: query
          schema:
            type: integer
        - name: bpm_max
          in: query
          schema:
            type: integer
        - name: key
          in: query
          schema:
            type: string
        - name: sort_by
          in: query
          schema:
            type: string
            enum: [title, artist, bpm, key, year, created_at]
            default: title
        - name: sort_order
          in: query
          schema:
            type: string
            enum: [asc, desc]
            default: desc
      responses:
        '200':
          description: Tracks retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TracksListResponse'

    post:
      tags:
        - Tracks
      summary: Add Track Manually
      description: Manually add a track to the library
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TrackCreate'
      responses:
        '201':
          description: Track created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Track'
  /tracks/{track_id}:
    get:
      tags:
        - Tracks
      summary: Get Track
      description: Get detailed information about a specific track
      parameters:
        - name: track_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Track retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Track'
        '404':
          description: Track not found

    put:
      tags:
        - Tracks
      summary: Update Track
      description: Update track metadata
      parameters:
        - name: track_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TrackUpdate'
      responses:
        '200':
          description: Track updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Track'

    delete:
      tags:
        - Tracks
      summary: Delete Track
      description: Remove a track from the library (does not delete the file)
      parameters:
        - name: track_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Track deleted successfully
  /tracks/bulk/delete:
    post:
      tags:
        - Tracks
      summary: Bulk Delete Tracks
      description: Delete multiple tracks from the library
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BulkDeleteTracksRequest'
      responses:
        '200':
          description: Tracks deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BulkDeleteTracksResponse'
...
  /playlists/{playlist_id}/tracks:
    post:
      tags:
        - Playlists
      summary: Add Tracks to Playlist
      description: Add one or more tracks to a playlist
      parameters:
        - name: playlist_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddTracksToPlaylistRequest'
      responses:
        '200':
          description: Tracks added successfully

    delete:
      tags:
        - Playlists
      summary: Remove Tracks from Playlist
      description: Remove tracks from a playlist
      parameters:
        - name: playlist_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RemoveTracksFromPlaylistRequest'
      responses:
        '200':
          description: Tracks removed successfully
...
components:
  schemas:
    Track:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        artist:
          type: string
        album:
          type: string
        year:
          type: integer
        genre:
          type: string
        mood:
          type: string
        bpm:
          type: integer
        key:
          type: string
        duration_seconds:
          type: integer
          description: Duration in seconds
        file_path:
          type: string
        file_size_bytes:
          type: integer
          description: File size in bytes
        file_format:
          type: string
          example: mp3
        bitrate_bps:
          type: integer
          description: Bitrate in bits per second
        sample_rate_hz:
          type: integer
          description: Sample rate in Hertz (Hz)
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        last_played:
          type: string
          format: date-time
        play_count:
          type: integer
        metadata_complete:
          type: boolean
    TrackCreate:
      type: object
      required:
        - file_path
      properties:
        file_path:
          type: string
        title:
          type: string
        artist:
          type: string
        album:
          type: string
        year:
          type: integer
        genre:
          type: string
        mood:
          type: string
        bpm:
          type: integer
        key:
          type: string
    TrackUpdate:
      type: object
      properties:
        title:
          type: string
        artist:
          type: string
        album:
          type: string
        year:
          type: integer
        genre:
          type: string
        mood:
          type: string
        bpm:
          type: integer
        key:
          type: string
...
    Pagination:
      type: object
      properties:
        page:
          type: integer
        size:
          type: integer
        total_pages:
          type: integer
        total_items:
          type: integer
        has_next:
          type: boolean
        has_previous:
          type: boolean
    TracksListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Track'
        pagination:
          $ref: '#/components/schemas/Pagination'
...
    BulkDeleteTracksRequest:
      type: object
      properties:
        track_ids:
          type: array
          items:
            type: string
            format: uuid
    BulkDeleteTracksResponse:
      type: object
      properties:
        deleted_count:
          type: integer
...
    AddTracksToPlaylistRequest:
      type: object
      properties:
        track_ids:
          type: array
          items:
            type: string
            format: uuid
    RemoveTracksFromPlaylistRequest:
      type: object
      properties:
        track_ids:
          type: array
          items:
            type: string
            format: uuid
...